requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.13
Django==4.1.4
django-braces==1.15.0
django-ckeditor==6.5.1
django-debug-toolbar==3.6.0
django-js-asset==2.0.0
django-redisboard==8.3.0
django-tinymce==3.6.1
idna==3.4
packaging==23.0
Pillow==9.2.0
pymemcache==3.5.2
pytz==2023.3
redis==4.3.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.7
urllib3==1.26.15
wrapt==1.15.0
